edit user detailed and user editor

+++ AnimalWebApp/src/main/webapp/controllers/AnimalsDetailedUserController.js
@@ -1,8 +1,15 @@
-//created by 41X
 var animalAppControllers = angular.module('AnimalsDetailedUserController', [ ]);
 
-animalApp.controller('AnimalsDetailedUserController', ['$scope', 'UserDataService', '$routeParams', '$window', 'AnimalsAdminValues',
-                                               function($scope, UserDataService, $routeParams, $window, AnimalsAdminValues) {
+animalApp.controller('AnimalsDetailedUserController', ['$scope', 'UserDataService', '$routeParams', '$window', 'UserAnimalsValues',
+                                                       'localStorageService', '$filter',
+                                               function($scope, UserDataService, $routeParams, $window, UserAnimalsValues, 
+                                            		   localStorageService, $filter) {
+	
+	if (localStorageService.get('userRole')!=="moderator" && 
+			localStorageService.get('userRole')!=="doctor" && 
+			localStorageService.get('userRole')!=="guest"){
+		$window.location.href = "#ua";
+	}
 	
 	 //initialize loading spinner
      var targetContent = document.getElementById('loading-block');
@@ -13,16 +20,16 @@ animalApp.controller('AnimalsDetailedUserController', ['$scope', 'UserDataServic
      $scope.currentLanguage = $window.localStorage.getItem('NG_TRANSLATE_LANG_KEY');
 
      var animalId = $routeParams.animalId;       //animal id
-     $scope.animal = AnimalsAdminValues.animal;  //animal
+     $scope.animal = UserAnimalsValues.animal;  //animal
      $scope.animalImage = undefined;	
 	
      UserDataService.getAnimal(animalId).then(
 			function(result){				
-				angular.copy(result, AnimalsAdminValues.animal);
+				angular.copy(result, UserAnimalsValues.animal);
 				$scope.animalImage = "resources/img/noimg.png";
-				if (AnimalsAdminValues.animal.image != undefined) {
-                    if (AnimalsAdminValues.animal.image.length > 0) {
-                        $scope.animalImage = AnimalsAdminValues.animal.image;
+				if (UserAnimalsValues.animal.image != undefined) {
+                    if (UserAnimalsValues.animal.image.length > 0) {
+                        $scope.animalImage = UserAnimalsValues.animal.image;
                     }
                 }
 				$scope.contentLoading--;
@@ -34,6 +41,10 @@ animalApp.controller('AnimalsDetailedUserController', ['$scope', 'UserDataServic
 		);
 	
      $scope.deleteAnimal = function() {
+    	 if (!confirm($filter('translate')("ANIMAL_DETAILED_CONFIRM_DELETE"))) {
+             return;
+         }
+    	 
     	 UserDataService.deleteAnimal($scope.animal.id)
             .then(function(data) {
                 $window.location.href = "#/ua/user/profile";
